p2p:remove broadcast V1 support
